backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
anthropic
openai
python-jose[cryptography]
passlib[bcrypt]>=1.7.4
python-multipart
stripe
aiosqlite
sqlalchemy[asyncio]
alembic
pydantic-settings
pydantic[email]
chardet
weasyprint
pandas
openpyxl
numpy
python-dateutil
jinja2
